+++ packages/taiko-client/driver/driver.go
@@ -106,7 +106,6 @@ func (d *Driver) InitFromConfig(ctx context.Context, cfg *Config) (err error) {
 		d.rpc,
 		d.state,
 		cfg.P2PSync,
-		cfg.P2PSyncTimeout,
 		cfg.BlobServerEndpoint,
 		latestSeenProposalCh,
 	); err != nil {